+++ packages/resource-time-grid/src/index.js
@@ -1,7 +1,6 @@
 import {btnTextDay, btnTextWeek, themeView, viewResources} from '@event-calendar/core';
 import TimeGrid from '@event-calendar/time-grid';
 import View from './View.svelte';
-import Auxiliary from './Auxiliary.svelte';
 
 export default {
 	createOptions(options) {
@@ -28,7 +27,6 @@ export default {
 		if (!('_times' in state)) {
 			TimeGrid.createStores(state);
 		}
-		state._auxiliary.update($_auxiliary => [...$_auxiliary, Auxiliary]);
 		if (!('_viewResources' in state)) {
 			state._viewResources = viewResources(state);
 		}