+++ packages/resource-timeline/src/Body.svelte
@@ -1,6 +1,7 @@
 <script>
     import {getContext} from 'svelte';
     import {toSeconds} from '@event-calendar/core';
+    import {getSlotTimeLimits} from './lib.js';
     import Days from './Days.svelte';
 
     let {_bodyEl, _headerEl, _sidebarEl, _dayTimes, _dayTimeLimits, _viewResources, _viewDates,
@@ -22,10 +23,8 @@
     }
 
     function scrollToTime() {
-        let slotTimeLimits = $_dayTimeLimits[$_viewDates[0].getTime()];
-        if (slotTimeLimits) {
-            el.scrollLeft = (toSeconds($scrollTime) - toSeconds(slotTimeLimits.min)) / toSeconds($slotDuration) * $slotWidth;
-        }
+        let slotTimeLimits = getSlotTimeLimits($_dayTimeLimits, $_viewDates[0]);
+        el.scrollLeft = (toSeconds($scrollTime) - toSeconds(slotTimeLimits.min)) / toSeconds($slotDuration) * $slotWidth;
     }
 </script>